# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/classes": {
      "post": {
        "operationId": "AdminClassesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      },
      "get": {
        "operationId": "AdminClassesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "ARCHIVED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      }
    },
    "/admin/classes/{id}": {
      "get": {
        "operationId": "AdminClassesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      },
      "patch": {
        "operationId": "AdminClassesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      }
    },
    "/admin/classes/{id}/archive": {
      "patch": {
        "operationId": "AdminClassesController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      }
    },
    "/admin/classes/{id}/unarchive": {
      "patch": {
        "operationId": "AdminClassesController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Classes"
        ]
      }
    },
    "/admin/invitations/template": {
      "get": {
        "operationId": "AdminInvitationsController_downloadTemplate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      }
    },
    "/admin/invitations": {
      "post": {
        "operationId": "AdminInvitationsController_createSingle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      },
      "get": {
        "operationId": "AdminInvitationsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "ACCEPTED",
                "EXPIRED",
                "REVOKED"
              ],
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      }
    },
    "/admin/invitations/bulk": {
      "post": {
        "operationId": "AdminInvitationsController_createBulk",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      }
    },
    "/admin/invitations/{id}/resend": {
      "post": {
        "operationId": "AdminInvitationsController_resend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      }
    },
    "/admin/invitations/{id}": {
      "delete": {
        "operationId": "AdminInvitationsController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Invitations"
        ]
      }
    },
    "/admin/students": {
      "get": {
        "operationId": "AdminStudentsController_list",
        "parameters": [
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Students"
        ]
      }
    },
    "/admin/students/{id}": {
      "get": {
        "operationId": "AdminStudentsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Students"
        ]
      },
      "patch": {
        "operationId": "AdminStudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Students"
        ]
      }
    },
    "/admin/students/{id}/progress": {
      "get": {
        "operationId": "AdminStudentsController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Students"
        ]
      }
    },
    "/admin/students/{id}/reset-password": {
      "post": {
        "operationId": "AdminStudentsController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin-access-token": []
          }
        ],
        "tags": [
          "Admin Students"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elearning Platform API",
    "description": "Backend NestJS platform kursus bahasa Jepang (single-tenant). Lihat docs/PRD.md dan docs/PLAN.md di repo untuk konteks produk lengkap.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateClassDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "targetLevelId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateClassDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "targetLevelId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "classId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "classId"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        }
      },
...
    }
  }
}